requirements.txt
langchain
langchain-community
sentence-transformers
spacy==3.7.2
networkx
cdlib
python-louvain
pypdf
numpy
scikit-learn
pyyaml
pandas
tqdm
streamlit==1.32.0
altair>=5.0.0
rank-bm25
langchain-text-splitters
groq
# Remove leidenalg if possible, or handle separately
# Remove ollama — won't work on cloud